==> 2. Merge Team Box Scores.ipynb <==
import numpy as np
import pandas as pd

ratings=pd.DataFrame()
advanced=pd.DataFrame()

for i in range(2000,2022):
    print(i)
    if i==2019:
        reg=pd.read_excel("Data/Euroleague/Team Box Score/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TEAMS',skiprows=1)
        cod=pd.read_excel("Data/Euroleague/Teams/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TM TRAD STATS',skiprows=2)

        reg=reg[reg["W/L"]!='TBP']
        cod=cod[cod["TEAM"].isna()==False]
        
        reg["TEAM"]=reg["TEAM"].apply(lambda x: str.upper(x))
        cod["TEAM"]=cod["TEAM"].apply(lambda x: str.upper(x))

        reg=reg.merge(cod[['TEAM','Code']],left_on='TEAM',right_on='TEAM',how='left')

        comb=reg
        comb['OR']=pd.to_numeric(comb['OR'])
        comb['OR.1']=pd.to_numeric(comb['OR.1'])
        comb['Pts']=pd.to_numeric(comb['Pts'])
        comb['PTS']=pd.to_numeric(comb['PTS'])

        comb['Home Poss']=comb['FGA']+0.44*comb['FTA']-1.07*(comb['OR']/(comb['OR']+comb['DR.1']))*(comb['FGA']-comb['FGM'])+comb['TO']
        comb['Away Poss']=comb['FGA.1']+0.44*comb['FTA.1']-1.07*(comb['OR.1']/(comb['OR.1']+comb['DR']))*(comb['FGA.1']-comb['FGM.1'])+comb['TO.1']
        comb['Poss']=(comb['Home Poss']+comb['Away Poss'])/2
        comb['Pace']=comb['Poss']/comb['Min']*200
        comb['Off Rat']=comb['Pts']/comb['Poss']*100
        comb['Def Rat']=comb['PTS']/comb['Poss']*100

        pv=pd.pivot_table(comb,index=['TEAM','Code'], values=['Poss','Pace','Off Rat','Def Rat'])
        pv2=pd.pivot_table(comb,index=['TEAM','Code'], values=['Away Poss','ST','BLK','2PA','Ast','FGM'],aggfunc='sum')
        pv2.rename(columns={'2PA':'2PTA.1'},inplace=True)

        pv['Year']=str(i)+"-"+str(i+1)
        pv2['Year']=str(i)+"-"+str(i+1)

        ratings=ratings.append(pv)
        advanced=advanced.append(pv2)
    elif i==2021:
        data=pd.read_excel("Data/Euroleague/Teams/2021-2022.xlsx",sheet_name='TEAMS_GbG',skiprows=2)
        
        data=data[data["MIN"]!='DNP']
        data=data[data["MIN"]!='TBP']
        data=data[data["MIN"].isna()==False]
        
        comb=data
        comb['Home Poss']=comb['FGA']+0.44*comb['FTA']-1.07*(comb['OR']/(comb['OR']+comb['DR.1']))*(comb['FGA']-comb['FGM'])+comb['TO']
        comb['Away Poss']=comb['FGA.1']+0.44*comb['FTA.1']-1.07*(comb['OR.1']/(comb['OR.1']+comb['DR']))*(comb['FGA.1']-comb['FGM.1'])+comb['TO.1']
        comb['Poss']=(comb['Home Poss']+comb['Away Poss'])/2
        comb['Pace']=comb['Poss']/comb['MIN']*200
        comb['Off Rat']=comb['PTS']/comb['Poss']*100
        comb['Def Rat']=comb['PTS.1']/comb['Poss']*100
        
        pv=pd.pivot_table(comb,index=['TM NAME','TM TAG'], values=['Poss','Pace','Off Rat','Def Rat'])
        pv2=pd.pivot_table(comb,index=['TM NAME','TM TAG'], values=['Away Poss','ST','BLK','2PTA.1','AST','FGM'],aggfunc='sum')
        pv2.rename(columns={'AST':'Ast'},inplace=True)
        pv['Year']=str(i)+"-"+str(i+1)
        pv2['Year']=str(i)+"-"+str(i+1)

        ratings=ratings.append(pv)
        advanced=advanced.append(pv2)

    else:
        reg=pd.read_excel("Data/Euroleague/Team Box Score/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TEAMS',skiprows=1)
        po=pd.read_excel("Data/Euroleague/Team Box Score/Playoff/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TEAMS',skiprows=1)
        cod=pd.read_excel("Data/Euroleague/Teams/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TM TRAD STATS',skiprows=2)

        reg=reg[reg["W/L"]!='TBP']
        po=po[po["W/L"]!='TBP']
        cod=cod[cod["TEAM"].isna()==False]
        
        reg["TEAM"]=reg["TEAM"].apply(lambda x: str.upper(x))
        po["TEAM"]=po["TEAM"].apply(lambda x: str.upper(x))
        cod["TEAM"]=cod["TEAM"].apply(lambda x: str.upper(x))

        reg=reg.merge(cod[['TEAM','Code']],left_on='TEAM',right_on='TEAM',how='left')
        po=po.merge(cod[['TEAM','Code']],left_on='TEAM',right_on='TEAM',how='left')

        comb=reg.append(po)
        comb['OR']=pd.to_numeric(comb['OR'])
        comb['OR.1']=pd.to_numeric(comb['OR.1'])
        comb['Pts']=pd.to_numeric(comb['Pts'])
        comb['PTS']=pd.to_numeric(comb['PTS'])

        comb['Home Poss']=comb['FGA']+0.44*comb['FTA']-1.07*(comb['OR']/(comb['OR']+comb['DR.1']))*(comb['FGA']-comb['FGM'])+comb['TO']
        comb['Away Poss']=comb['FGA.1']+0.44*comb['FTA.1']-1.07*(comb['OR.1']/(comb['OR.1']+comb['DR']))*(comb['FGA.1']-comb['FGM.1'])+comb['TO.1']
        comb['Poss']=(comb['Home Poss']+comb['Away Poss'])/2
        comb['Pace']=comb['Poss']/comb['Min']*200
        comb['Off Rat']=comb['Pts']/comb['Poss']*100
        comb['Def Rat']=comb['PTS']/comb['Poss']*100

        pv=pd.pivot_table(comb,index=['TEAM','Code'], values=['Poss','Pace','Off Rat','Def Rat'])
        pv2=pd.pivot_table(comb,index=['TEAM','Code'], values=['Away Poss','ST','BLK','2PA','Ast','FGM'],aggfunc='sum')
        pv2.rename(columns={'2PA':'2PTA.1'},inplace=True)

        pv['Year']=str(i)+"-"+str(i+1)
        pv2['Year']=str(i)+"-"+str(i+1)

        ratings=ratings.append(pv)
        advanced=advanced.append(pv2)

ratings['Net Rat']=ratings['Off Rat']-ratings['Def Rat']

advanced['Ast%']=advanced['Ast']/advanced['FGM']*100
advanced['ST%']=advanced['ST']/advanced['Away Poss']*100
advanced['BLK%']=advanced['BLK']/advanced['2PTA.1']*100

ratings.to_csv('euroleague_ratings.csv')
advanced.to_csv('euroleague_advanced.csv')
