==> 3. Merge NBA BBRef.ipynb <==
import numpy as np
import pandas as pd

nba_players_adv=pd.DataFrame()

for i in range(2000,2022):
    data=pd.read_excel("Data/NBA/players_advanced.xlsx",sheet_name=str(i)+"-"+str(i+1))
    data['year']=str(i)+"-"+str(i+1)
    nba_players_adv=nba_players_adv.append(data)

nba_players_adv.drop(columns='Rk',inplace=True)
nba_players_adv.to_csv('nba_players_adv.csv',index=False)

nba_teams_adv=pd.DataFrame()

for i in range(2000,2022):
    data=pd.read_excel("Data/NBA/teams_advanced.xlsx",sheet_name=str(i)+"-"+str(i+1))
    data['year']=str(i)+"-"+str(i+1)
    nba_teams_adv=nba_teams_adv.append(data)

nba_teams_adv.drop(columns='Rk',inplace=True)
nba_teams_adv.to_csv('nba_teams_adv.csv',index=False)

nba_players_per100=pd.DataFrame()

for i in range(2000,2022):
    data=pd.read_excel("Data/NBA/players_per100.xlsx",sheet_name=str(i)+"-"+str(i+1))
    data['year']=str(i)+"-"+str(i+1)
    nba_players_per100=nba_players_per100.append(data)

nba_players_per100.drop(columns=['Rk','Unnamed: 29'],inplace=True)

nba_players_basic=pd.DataFrame()

for i in range(2000,2022):
    data=pd.read_excel("Data/NBA/players_basic.xlsx",sheet_name=str(i)+"-"+str(i+1))
    data['year']=str(i)+"-"+str(i+1)
    nba_players_basic=nba_players_basic.append(data)

nba_players_basic.drop(columns='Rk',inplace=True)

nba_players_basic['id']=nba_players_basic['Player']+nba_players_basic['year']+nba_players_basic['Tm']+nba_players_basic['Pos']
nba_players_per100['id']=nba_players_per100['Player']+nba_players_per100['year']+nba_players_per100['Tm']+nba_players_per100['Pos']
nba_players_per100.rename(columns={'FG':'FGper100','FGA':'FGAper100','3P':'3Pper100','3PA':'3PAper100','2P':'2Pper100','2PA':'2PAper100','FT':'FTper100','FTA':'FTAper100','ORB':'ORBper100','DRB':'DRBper100','TRB':'TRBper100','AST':'ASTper100','STL':'STLper100','BLK':'BLKper100','TOV':'TOVper100','PF':'PFper100','PTS':'PTSper100','ORtg':'ORtgper100','DRtg':'DRtgper100'},inplace=True)
nba_players_per100=nba_players_per100[['id','FGper100','FGAper100','3Pper100','3PAper100','2Pper100','2PAper100','FTper100','FTAper100','ORBper100','DRBper100','TRBper100','ASTper100','STLper100','BLKper100','TOVper100','PFper100','PTSper100','ORtgper100','DRtgper100']]

nba_players_basic_per_100=nba_players_basic.merge(nba_players_per100,right_on='id',left_on='id',how='left')
nba_players_basic_per_100['id']=nba_players_basic_per_100['Player']+nba_players_basic_per_100['year']
nba_players_basic['id']=nba_players_basic['Player']+nba_players_basic['year']
nba_players_per100['id']=nba_players_per100['Player']+nba_players_per100['year']

nba_players_basic.to_csv('nba_players_basic.csv',index=False)
nba_players_per100.to_csv('nba_players_per100.csv',index=False)
nba_players_basic_per_100.to_csv('nba_players_basic_per_100.csv',index=False)

nba_players_shooting=pd.DataFrame()

for i in range(2000,2022):
    data=pd.read_excel("Data/NBA/players_shooting.xlsx",sheet_name=str(i)+"-"+str(i+1))
    data['year']=str(i)+"-"+str(i+1)
    nba_players_shooting=nba_players_shooting.append(data)

nba_players_shooting.drop(columns='Rk',inplace=True)
nba_players_shooting.to_csv('nba_players_shooting.csv',index=False)

recruiting_ranks=pd.DataFrame()

for i in range(2000,2022):
    data=pd.read_excel("Data/NBA/recruiting_ranks.xlsx",sheet_name=str(i))
    data['year']=str(i)
    recruiting_ranks=recruiting_ranks.append(data)

recruiting_ranks.to_csv('recruiting_ranks.csv',index=False)
