==> 4. Code Joins.ipynb <==
import numpy as np
import pandas as pd

teams_esp=pd.read_csv('team_esp.csv')
teams_it=pd.read_csv('team_it.csv')
teams_tur=pd.read_csv('team_tur.csv')
teams_gr=pd.read_csv('team_gr.csv')
teams_fr=pd.read_csv('team_fr.csv')
teams_rus=pd.read_csv('team_rus.csv')
teams_adr=pd.read_csv('team_adr.csv')

names=pd.read_excel('Names.xlsx')

names['Local Name']=names['Local Name'].apply(lambda x: x.upper())
names=names.sort_values(by='Local Name')

teams_esp['Team']=teams_esp['Team'].apply(lambda x: x.upper())
teams_esp=teams_esp.sort_values(by='Team')
teams_esp=teams_esp.reset_index(drop=True)
names_tba=names[names['Local Country']=='ESP']['Code TBU']
names_tba=names_tba.reset_index(drop=True)
teams_esp.insert(loc=1,column='Code TBU', value=names_tba)
teams_esp['New id']=teams_esp['Code TBU']+teams_esp['Year']
teams_esp.to_csv('team_esp.csv',index=False)

teams_it['Team']=teams_it['Team'].apply(lambda x: x.upper())
teams_it=teams_it.sort_values(by='Team')
teams_it=teams_it.reset_index(drop=True)
names_tba=names[names['Local Country']=='IT']['Code TBU']
names_tba=names_tba.reset_index(drop=True)
teams_it.insert(loc=1,column='Code TBU', value=names_tba)
teams_it['New id']=teams_it['Code TBU']+teams_it['Year']
teams_it.to_csv('team_it.csv',index=False)

teams_tur['Team']=teams_tur['Team'].apply(lambda x: x.upper())
teams_tur=teams_tur.sort_values(by='Team')
teams_tur=teams_tur.reset_index(drop=True)
names_tba=names[names['Local Country']=='TUR']['Code TBU']
names_tba=names_tba.reset_index(drop=True)
teams_tur.insert(loc=1,column='Code TBU', value=names_tba)
teams_tur['New id']=teams_tur['Code TBU']+teams_tur['Year']
teams_tur.to_csv('team_tur.csv',index=False)

teams_gr['Team']=teams_gr['Team'].apply(lambda x: x.upper())
teams_gr=teams_gr.sort_values(by='Team')
teams_gr=teams_gr.reset_index(drop=True)
names_tba=names[names['Local Country']=='GR']['Code TBU']
names_tba=names_tba.reset_index(drop=True)
teams_gr.insert(loc=1,column='Code TBU', value=names_tba)
teams_gr['New id']=teams_gr['Code TBU']+teams_gr['Year']
teams_gr.to_csv('team_gr.csv',index=False)

teams_fr['Team']=teams_fr['Team'].apply(lambda x: x.upper())
teams_fr=teams_fr.sort_values(by='Team')
teams_fr=teams_fr.reset_index(drop=True)
names_tba=names[names['Local Country']=='FR']['Code TBU']
names_tba=names_tba.reset_index(drop=True)
teams_fr.insert(loc=1,column='Code TBU', value=names_tba)
teams_fr['New id']=teams_fr['Code TBU']+teams_fr['Year']
teams_fr.to_csv('team_fr.csv',index=False)

teams_adr['Team']=teams_adr['Team'].apply(lambda x: x.upper())
teams_adr=teams_adr.sort_values(by='Team')
teams_adr=teams_adr.reset_index(drop=True)
names_tba=names[names['Local Country']=='ADR']['Code TBU']
names_tba=names_tba.reset_index(drop=True)
teams_adr.insert(loc=1,column='Code TBU', value=names_tba)
teams_adr['New id']=teams_adr['Code TBU']+teams_adr['Year']
teams_adr.to_csv('team_adr.csv',index=False)

teams_rus['Team']=teams_rus['Team'].apply(lambda x: x.upper())
teams_rus=teams_rus.sort_values(by='Team')
teams_rus=teams_rus.reset_index(drop=True)
names_tba=names[names['Local Country']=='RUS']['Code TBU']
names_tba=names_tba.reset_index(drop=True)
teams_rus.insert(loc=1,column='Code TBU', value=names_tba)
teams_rus['New id']=teams_rus['Code TBU']+teams_rus['Year']
teams_rus.to_csv('team_rus.csv',index=False)

players_esp=pd.read_csv('players_esp.csv')
players_it=pd.read_csv('players_it.csv')
players_tur=pd.read_csv('players_tur.csv')
players_gr=pd.read_csv('players_gr.csv')
players_fr=pd.read_csv('players_fr.csv')
players_rus=pd.read_csv('players_rus.csv')
players_adr=pd.read_csv('players_adr.csv')

names=pd.read_excel('Names.xlsx')

names_tba=names[names['Country']=='ESP'][['Team Code','Player Code']]
players_esp=players_esp.merge(names_tba, left_on='Team',right_on='Player Code',how='inner')
players_esp.drop(columns='Player Code',inplace=True)
players_esp['New id']=players_esp['Team Code']+players_esp['Year']
players_esp.to_csv('players_esp.csv',index=False)

names_tba=names[names['Country']=='IT'][['Team Code','Player Code']]
players_it=players_it.merge(names_tba, left_on='Team',right_on='Player Code',how='inner')
players_it.drop(columns='Player Code',inplace=True)
players_it['New id']=players_it['Team Code']+players_it['Year']
players_it.to_csv('players_it.csv',index=False)

names_tba=names[names['Country']=='TUR'][['Team Code','Player Code']]
players_tur=players_tur.merge(names_tba, left_on='Team',right_on='Player Code',how='inner')
players_tur.drop(columns='Player Code',inplace=True)
players_tur['New id']=players_tur['Team Code']+players_tur['Year']
players_tur.to_csv('players_tur.csv',index=False)

names_tba=names[names['Country']=='GR'][['Team Code','Player Code']]
players_gr=players_gr.merge(names_tba, left_on='Team',right_on='Player Code',how='inner')
players_gr.drop(columns='Player Code',inplace=True)
players_gr['New id']=players_gr['Team Code']+players_gr['Year']
players_gr.to_csv('players_gr.csv',index=False)

names_tba=names[names['Country']=='FR'][['Team Code','Player Code']]
players_fr=players_fr.merge(names_tba, left_on='Team',right_on='Player Code',how='inner')
players_fr.drop(columns='Player Code',inplace=True)
players_fr['New id']=players_fr['Team Code']+players_fr['Year']
players_fr.to_csv('players_fr.csv',index=False)

names_tba=names[names['Country']=='RUS'][['Team Code','Player Code']]
players_rus=players_rus.merge(names_tba, left_on='Team',right_on='Player Code',how='inner')
players_rus.drop(columns='Player Code',inplace=True)
players_rus['New id']=players_rus['Team Code']+players_rus['Year']
players_rus.to_csv('players_rus.csv',index=False)

names_tba=names[names['Country']=='ADR'][['Team Code','Player Code']]
players_adr=players_adr.merge(names_tba, left_on='Team',right_on='Player Code',how='inner')
players_adr.drop(columns='Player Code',inplace=True)
players_adr['New id']=players_adr['Team Code']+players_adr['Year']
players_adr.to_csv('players_adr.csv',index=False)
