==> 5. Euroleague Player Totals.ipynb <==
import numpy as np
import pandas as pd

players=pd.DataFrame()

for i in range(2000,2022):
    if i==2019:
        reg=pd.read_excel("Data/Euroleague/Players/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='PLAYERS TRAD STATS',skiprows=1)
    
        reg=reg[reg["NAME"].isna()==False]
        reg=reg[reg['GP']!='DNP']

        reg=reg[['NAME','ROLE','TEAM','GP','Min/G','FGA/G','3PM/G','FTA/G','OR/G','DR/G','TR/G', 'Ast/G','ST/G','BLK/G','TO/G','PF/G','PTS/G']]

        comb=reg

        comb['Min']=comb['Min/G']*comb['GP']
        comb['FGA']=comb['FGA/G']*comb['GP']
        comb['3PM']=comb['3PM/G']*comb['GP']
        comb['FTA']=comb['FTA/G']*comb['GP']
        comb['OR']=comb['OR/G']*comb['GP']
        comb['DR']=comb['DR/G']*comb['GP']
        comb['TR']=comb['TR/G']*comb['GP']
        comb['AST']=comb['Ast/G']*comb['GP']
        comb['ST']=comb['ST/G']*comb['GP']
        comb['BLK']=comb['BLK/G']*comb['GP']
        comb['TO']=comb['TO/G']*comb['GP']
        comb['PF']=comb['PF/G']*comb['GP']
        comb['PTS']=comb['PTS/G']*comb['GP']

        comb['year']=str(i)+"-"+str(i+1)
        comb['id']=comb['NAME']+comb['year']

        comb_totals=comb.drop(columns=['Min/G','FGA/G','3PM/G','FTA/G','OR/G','DR/G','TR/G', 'Ast/G','ST/G','BLK/G','TO/G','PF/G','PTS/G'])

        totals=comb_totals.groupby(['id'])[['GP', 'Min', 'FGA', '3PM', 'FTA', 'OR', 'DR', 'TR', 'AST', 'ST', 'BLK', 'TO', 'PF', 'PTS']].sum()

        totals=totals.merge(comb_totals[['id','NAME', 'ROLE', 'TEAM']],left_on='id',right_on='id',how='left')

        totals=totals[['id','NAME', 'ROLE', 'TEAM', 'GP','Min', 'FGA', '3PM', 'FTA', 'OR', 'DR', 'TR', 'AST', 'ST', 'BLK','TO', 'PF', 'PTS']]

        totals.drop(totals[totals['NAME']=='AVERAGE'].index,inplace=True)

        totals['Min']=totals['Min'].apply(lambda x: round(x,0))

        totals=totals.drop_duplicates()

        players=players.append(totals)
        
    elif i==2021:
        data=pd.read_excel("Data/Euroleague/Players/2021-2022.xlsx",sheet_name='PLAYERS_STATS',skiprows=9)

        data=data[data["NAME"].isna()==False]
        data=data[data['MIN']!='DNP']

        data=data[['NAME','ROLE','TEAM TAG','GP','MIN','FGA','3PTM','FTA','OR','DR','TR', 'AST','ST','BLK','TO','PF','PTS']]

        data['year']='2021-2022'

        data['id']=data['NAME']+data['year']

        data=data[['id','NAME', 'ROLE', 'TEAM TAG', 'GP','MIN', 'FGA', '3PTM', 'FTA', 'OR', 'DR', 'TR', 'AST', 'ST', 'BLK','TO', 'PF', 'PTS']]

        data['MIN']=data['MIN'].apply(lambda x: round(x,0))

        data = data.rename(columns={'TEAM TAG': 'TEAM', 'MIN': 'Min', '3PTM': '3PM'})
        
        players=players.append(data)

    else:
        reg=pd.read_excel("Data/Euroleague/Players/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='PLAYERS TRAD STATS',skiprows=1)
        po=pd.read_excel("Data/Euroleague/Players/Playoffs/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='PLAYERS TRAD STATS',skiprows=1)

        reg=reg[reg["NAME"].isna()==False]
        reg=reg[reg['GP']!='DNP']
        po=po[po["TEAM"].isna()==False]
        po=po[po['GP']!='DNP']

        reg=reg[['NAME','ROLE','TEAM','GP','Min/G','FGA/G','3PM/G','FTA/G','OR/G','DR/G','TR/G', 'Ast/G','ST/G','BLK/G','TO/G','PF/G','PTS/G']]
        po=po[['NAME','ROLE','TEAM','GP','Min/G','FGA/G','3PM/G','FTA/G','OR/G','DR/G','TR/G', 'Ast/G','ST/G','BLK/G','TO/G','PF/G','PTS/G']]

        comb=reg.append(po)

        comb['Min']=comb['Min/G']*comb['GP']
        comb['FGA']=comb['FGA/G']*comb['GP']
        comb['3PM']=comb['3PM/G']*comb['GP']
        comb['FTA']=comb['FTA/G']*comb['GP']
        comb['OR']=comb['OR/G']*comb['GP']
        comb['DR']=comb['DR/G']*comb['GP']
        comb['TR']=comb['TR/G']*comb['GP']
        comb['AST']=comb['Ast/G']*comb['GP']
        comb['ST']=comb['ST/G']*comb['GP']
        comb['BLK']=comb['BLK/G']*comb['GP']
        comb['TO']=comb['TO/G']*comb['GP']
        comb['PF']=comb['PF/G']*comb['GP']
        comb['PTS']=comb['PTS/G']*comb['GP']

        comb['year']=str(i)+"-"+str(i+1)
        comb['id']=comb['NAME']+comb['year']

        comb_totals=comb.drop(columns=['Min/G','FGA/G','3PM/G','FTA/G','OR/G','DR/G','TR/G', 'Ast/G','ST/G','BLK/G','TO/G','PF/G','PTS/G'])

        totals=comb_totals.groupby(['id'])[['GP', 'Min', 'FGA', '3PM', 'FTA', 'OR', 'DR', 'TR', 'AST', 'ST', 'BLK', 'TO', 'PF', 'PTS']].sum()

        totals=totals.merge(comb_totals[['id','NAME', 'ROLE', 'TEAM']],left_on='id',right_on='id',how='left')

        totals=totals[['id','NAME', 'ROLE', 'TEAM', 'GP','Min', 'FGA', '3PM', 'FTA', 'OR', 'DR', 'TR', 'AST', 'ST', 'BLK','TO', 'PF', 'PTS']]

        totals.drop(totals[totals['NAME']=='AVERAGE'].index,inplace=True)

        totals['Min']=totals['Min'].apply(lambda x: round(x,0))

        totals=totals.drop_duplicates()

        players=players.append(totals)

players

players.to_csv('euroleague_players.csv',index=False)
