==> 6. Euroleague Team Totals.ipynb <==
import numpy as np
import pandas as pd

teams=pd.DataFrame()

for i in range(2000,2022):
    if i==2019:
        reg=pd.read_excel("Data/Euroleague/Teams/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TM TRAD STATS',skiprows=2)
        
        reg=reg[reg["TEAM"].isna()==False]
        reg=reg[reg["TEAM"]!='AVERAGE']


        reg=reg[['TEAM','Code','GP','Min','Pts','FGA','FTA','TR', 'Ast','ST','BLK','PF']]

        comb=reg

        comb['Min']=comb['Min']*comb['GP']
        comb['FGA']=comb['FGA']*comb['GP']
        comb['FTA']=comb['FTA']*comb['GP']
        comb['TR']=comb['TR']*comb['GP']
        comb['Ast']=comb['Ast']*comb['GP']
        comb['ST']=comb['ST']*comb['GP']
        comb['BLK']=comb['BLK']*comb['GP']
        comb['PF']=comb['PF']*comb['GP']
        comb['Pts']=comb['Pts']*comb['GP']
        
        comb['year']=str(i)+"-"+str(i+1)
        comb['id']=comb['Code']+comb['year']

        totals_gp=comb.groupby(['id'])[['GP']].sum()
        totals_rest=comb.groupby(['id'])[['Min', 'FGA', 'FTA', 'TR', 'Ast', 'ST', 'BLK', 'PF', 'Pts']].sum()
        totals=totals_gp.join(totals_rest)

        totals=totals.merge(comb[['id', 'TEAM','Code']],left_on='id',right_on='id',how='left')
        totals=totals[['id','TEAM','Code', 'GP','Min', 'FGA', 'FTA', 'TR', 'Ast', 'ST', 'BLK', 'PF', 'Pts']]

        totals['Min']=totals['Min'].apply(lambda x: round(x,0))
        totals=totals.drop_duplicates()
        teams=teams.append(totals)
        
    elif i==2021:
        data=pd.read_excel("Data/Euroleague/Teams/2021-2022.xlsx",sheet_name='TEAMS_STATS',skiprows=9)

        data=data[data["TEAM"].isna()==False]
        
        data=data[['TEAM','TM TAG','GP','MIN','PTS','FGA','FTA','TR', 'AST','ST','BLK','PF']]
        
        data['year']='2021-2022'
        data['id']=data['TM TAG']+data['year']
        
        data=data[['id','TEAM', 'TM TAG', 'GP','MIN','FGA', 'FTA', 'TR', 'AST', 'ST', 'BLK', 'PF', 'PTS']]
        data = data.rename(columns={'TM TAG': 'Code', 'MIN': 'Min', 'PTS': 'Pts', 'AST':'Ast'})
        data['Min']=data['Min'].apply(lambda x: round(x,0))

        teams=teams.append(data)

    else:
        reg=pd.read_excel("Data/Euroleague/Teams/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TM TRAD STATS',skiprows=2)
        po=pd.read_excel("Data/Euroleague/Teams/Playoffs/"+str(i)+"-"+str(i+1)+".xlsx",sheet_name='TM TRAD STATS',skiprows=2)
        
        reg=reg[reg["TEAM"].isna()==False]
        reg=reg[reg["TEAM"]!='AVERAGE']
        po=po[po["TEAM"].isna()==False]
        po=po[po['GP']!=0]
        po=po[po["TEAM"]!='AVERAGE']

        reg=reg[['TEAM','Code','GP','Min','Pts','FGA','FTA','TR', 'Ast','ST','BLK','PF']]
        po=po[['TEAM','Code','GP','Min','Pts','FGA','FTA','TR', 'Ast','ST','BLK','PF']]

        comb=reg.append(po)

        comb['Min']=comb['Min']*comb['GP']
        comb['FGA']=comb['FGA']*comb['GP']
        comb['FTA']=comb['FTA']*comb['GP']
        comb['TR']=comb['TR']*comb['GP']
        comb['Ast']=comb['Ast']*comb['GP']
        comb['ST']=comb['ST']*comb['GP']
        comb['BLK']=comb['BLK']*comb['GP']
        comb['PF']=comb['PF']*comb['GP']
        comb['Pts']=comb['Pts']*comb['GP']
        
        comb['year']=str(i)+"-"+str(i+1)
        comb['id']=comb['Code']+comb['year']

        totals_gp=comb.groupby(['id'])[['GP']].sum()
        totals_rest=comb.groupby(['id'])[['Min', 'FGA', 'FTA', 'TR', 'Ast', 'ST', 'BLK', 'PF', 'Pts']].sum()
        totals=totals_gp.join(totals_rest)

        totals=totals.merge(comb[['id', 'TEAM','Code']],left_on='id',right_on='id',how='left')
        totals=totals[['id','TEAM','Code', 'GP','Min', 'FGA', 'FTA', 'TR', 'Ast', 'ST', 'BLK', 'PF', 'Pts']]
        totals['Min']=totals['Min'].apply(lambda x: round(x,0))

        totals=totals.drop_duplicates()
        teams=teams.append(totals)
    print(i)

teams.to_csv('euroleague_teams.csv',index=False)
