==> 7. NBA Advanced.ipynb <==
import numpy as np
import pandas as pd

data_21_22=pd.read_csv("Data/NBA/Team Box Scores/Games_data_21-22.csv")

data_21_22_home=data_21_22.drop(data_21_22[(data_21_22.index % 2)!=0].index)
data_21_22_away=data_21_22.drop(data_21_22[(data_21_22.index % 2)==0].index)

data_21_22_home.reset_index(drop=True,inplace=True)
data_21_22_away.reset_index(drop=True,inplace=True)

data_21_22_home_new=data_21_22_home.join(data_21_22_away,rsuffix='_Opp')
data_21_22_away_new=data_21_22_away.join(data_21_22_home,rsuffix='_Opp')

data_21_22=data_21_22_home_new.append(data_21_22_away_new)

data_21_22['Year']='2021-2022'
data_21_22['Away Poss']=data_21_22['FGA_Opp']+0.44*data_21_22['FTA_Opp']-1.07*(data_21_22['OREB_Opp']/(data_21_22['OREB_Opp']+data_21_22['DREB']))*(data_21_22['FGA_Opp']-data_21_22['FGM_Opp'])+data_21_22['TOV_Opp']

data_21_22['id']=data_21_22['TEAM_ABBREVIATION']+data_21_22['Year']
data_21_22['2PA']=data_21_22['FGA']-data_21_22['FG3A']
data_21_22_pv=pd.pivot_table(data_21_22,index='id', values=['Away Poss','STL','BLK','2PA','AST','FGM'],aggfunc='sum')

data_00_21=pd.read_csv("Data/NBA/Team Box Scores/Games_data_00-21.csv")

data_00_21=data_00_21[data_00_21['SEASON_ID']>=22000]

data_00_21=data_00_21.drop_duplicates(subset='GAME_ID')

data_00_21['Year']=data_00_21['SEASON_ID'].apply(lambda x: str(x-20000)+'-'+str(x-20000+1))

data_00_21_home=data_00_21.copy()
data_00_21_away=data_00_21.copy()

data_00_21_home['Away Poss']=data_00_21_home['FGA_AWAY']+0.44*data_00_21_home['FTA_AWAY']-1.07*(data_00_21_home['OREB_AWAY']/(data_00_21_home['OREB_AWAY']+data_00_21_home['DREB_HOME']))*(data_00_21_home['FGA_AWAY']-data_00_21_home['FGM_AWAY'])+data_00_21_home['TOV_AWAY']
data_00_21_away['Away Poss']=data_00_21_away['FGA_HOME']+0.44*data_00_21_away['FTA_HOME']-1.07*(data_00_21_away['OREB_HOME']/(data_00_21_away['OREB_HOME']+data_00_21_away['DREB_AWAY']))*(data_00_21_away['FGA_HOME']-data_00_21_away['FGM_HOME'])+data_00_21_away['TOV_HOME']

data_00_21_home['id']=data_00_21_home['TEAM_ABBREVIATION_HOME']+data_00_21_home['Year']
data_00_21_home['2PA']=data_00_21_home['FGA_HOME']-data_00_21_home['FG3A_HOME']
data_00_21_home['STL']=data_00_21_home['STL_HOME']
data_00_21_home['BLK']=data_00_21_home['BLK_HOME']
data_00_21_home['AST']=data_00_21_home['AST_HOME']
data_00_21_home['FGM']=data_00_21_home['FGM_HOME']

data_00_21_away['id']=data_00_21_away['TEAM_ABBREVIATION_AWAY']+data_00_21_away['Year']
data_00_21_away['2PA']=data_00_21_away['FGA_AWAY']-data_00_21_away['FG3A_AWAY']
data_00_21_away['STL']=data_00_21_away['STL_AWAY']
data_00_21_away['BLK']=data_00_21_away['BLK_AWAY']
data_00_21_away['AST']=data_00_21_away['AST_AWAY']
data_00_21_away['FGM']=data_00_21_away['FGM_AWAY']

data_00_21=data_00_21_home.append(data_00_21_away)

data_00_21_pv=pd.pivot_table(data_00_21,index='id', values=['Away Poss','STL','BLK','2PA','AST','FGM'],aggfunc='sum')

data=data_00_21_pv.append(data_21_22_pv)

data['Year']=data.index
data['Year']=data['Year'].apply(lambda x: x[3:])

data['Ast%']=data['AST']/data['FGM']*100
data['ST%']=data['STL']/data['Away Poss']*100
data['BLK%']=data['BLK']/data['2PA']*100

data.to_csv('nba_team_adv_for_comparison.csv')
