==> 8. League Comparison.ipynb <==
import numpy as np
import pandas as pd

nba_pct=pd.read_csv('nba_team_adv_for_comparison.csv')
euro_pct=pd.read_csv('euroleague_advanced.csv')
nba_ratings=pd.read_csv('nba_teams_adv.csv')
euro_ratings=pd.read_csv('euroleague_ratings.csv')

nba_pct_mean=pd.pivot_table(nba_pct,values=['Ast%','ST%','BLK%'], index=['Year'],aggfunc='mean')
euro_pct_mean=pd.pivot_table(euro_pct,values=['Ast%','ST%','BLK%'], index=['Year'],aggfunc='mean')

nba_ratings_mean=pd.pivot_table(nba_ratings,values=['ORtg/A','DRtg/A'],index=['year'],aggfunc='mean')
euro_ratings_mean=pd.pivot_table(euro_ratings,values=['Off Rat','Def Rat'],index=['Year'],aggfunc='mean')

nba=nba_pct_mean.join(nba_ratings_mean)
euro=euro_pct_mean.join(euro_ratings_mean)

diffs=pd.DataFrame(index=nba.index)

diffs['Ast%']=nba['Ast%']-euro['Ast%']
diffs['BLK%']=nba['BLK%']-euro['BLK%']
diffs['ST%']=nba['ST%']-euro['ST%']
diffs['Off Rat']=nba['ORtg/A']-euro['Off Rat']
diffs['Def Rat']=nba['DRtg/A']-euro['Def Rat']

diffs
