==> 11. Player Height Scrape.ipynb <==
import numpy as np
import pandas as pd
from urllib.request import urlopen
from bs4 import BeautifulSoup
import re

lst=[]
for i in range(2001,2023):
    url = 'https://www.basketball-reference.com/leagues/NBA_'+str(i)+'_totals.html'
    html = urlopen(url)
    soup = BeautifulSoup(html, features="lxml")
    table =soup.select('table')[0]
    rows = table.find_all('tr')
    for row in rows:
        lst.append(row.find_all('a', href=True))

links=[]
names=[]
for i in range(len(lst)):
    try:
        s=str(lst[i][0])
        names.append(re.findall('>([^"]*)<', s))
        links.append(re.findall('"([^"]*)"', s))
    except:
        pass

df=pd.DataFrame(names,columns=['names'])

df['link']=pd.DataFrame(links,columns=['link'])

df.drop_duplicates(inplace=True)

df.reset_index(drop=True,inplace=True)

height=pd.DataFrame(index=df['link'],columns=['height'])
for i in df['link']:
    url = 'https://www.basketball-reference.com'+i
    html = urlopen(url)
    soup = BeautifulSoup(html, features="lxml")
    s=soup.find(string=re.compile("cm,"))
    loc=s.find('cm,')
    height.loc[i,'height']=s[loc-3:loc]

height['height'] = height['height'].apply(pd.to_numeric, errors='coerce')
height.reset_index(drop=True, inplace=True)
df['height']=height['height']

df.to_csv('height.csv',index=False)

# Other bio stats

other_stats=pd.DataFrame(index=df['link'],columns=['birth_year','college','recruiting_rank','draft','exp'])
for i in df['link']:
    url = 'https://www.basketball-reference.com'+i
    html = urlopen(url)
    soup = BeautifulSoup(html, features="lxml")
    birth_year=soup.find(id='necro-birth')
    birth_year=str(birth_year)
    loc=birth_year.find('data-birth')
    other_stats.loc[i,'birth_year']=birth_year[loc+12:loc+16]


    col=soup.select("a[href*=colleges]")
    try:
        college=re.findall('>([^"]*)<', str(col[1]))[0]
        other_stats.loc[i,'college']=college
    except:
        other_stats.loc[i,'college']='no college'

    s=str(soup.find(id='meta'))
    loc=s.find('Recruiting Rank')
    if loc==-1:
        other_stats.loc[i,'recruiting_rank']='no rank'
    else:
        rec=s[loc:]
        loc=rec.find('(')
        other_stats.loc[i,'recruiting_rank']=rec[loc+1:loc+2]

    loc=s.find(' round')
    if loc==-1:
        other_stats.loc[i,'draft']='undrafted'
    else:
        draft=s[loc:]
        loc=draft.find('overall')
        other_stats.loc[i,'draft']=int(draft[loc-5:loc-3].strip())

   
    loc=s.find('Experience')
    exp=s[loc:]
    loc=exp.find('year')
    if exp[loc-3:loc-1]!='':
        other_stats.loc[i,'exp']=int(exp[loc-3:loc-1])
    else:
        loc=s.find('Career Length')
        exp=s[loc:]
        loc=exp.find('year')
        other_stats.loc[i,'exp']=int(exp[loc-3:loc-1])

other_stats

other_stats.isnull().sum()

other_stats=pd.concat([df,other_stats],axis=1)

other_stats.to_csv('nba_other_stats.csv',index=False)

leagues = {1: ['4','Spanish-ACB'],
           2: ['6','Italian-Lega-Basket-Serie-A'],
           3: ['7', 'Turkish-BSL'],
           4: ['8', 'Greek-HEBA-A1'],
           5: ['12', 'French-Jeep-Elite'],
           6: ['18', 'Adriatic-League-Liga-ABA'],
           7: ['35','VTB-United-League']
          }

def get_player_heights(n):
    print(leagues[n][1])
    lst= []
    for i in range(2004,2023):
        for j in range(1,5):
            test=0
            while test==0:
                try:
                    url = 'https://basketball.realgm.com/international/league/'+leagues[n][0]+'/'+leagues[n][1]+'/stats/'+str(i)+'/Totals/All/All/points/All/desc/'+str(j)
                    html = urlopen(url)
                    soup = BeautifulSoup(html, features="lxml")
                    test=1
                except:
                    continue
                break
            try:
                table =soup.select('table')[0]
                rows = table.find_all('tr')
                for row in rows[1:]:
                    x=row.find_all('a', href=True)
                    lst.append(row.find_all('a', href=True))
            except:
                continue
    return lst

esp_height=get_player_heights(1)

it_height=get_player_heights(2)

tur_height=get_player_heights(3)

gr_height=get_player_heights(4)

fr_height=get_player_heights(5)

adr_height=get_player_heights(6)

rus_height=get_player_heights(7)

eur_lst=esp_height+it_height+tur_height+gr_height+fr_height+adr_height+rus_height

eur_links=[]
eur_names=[]
for i in range(len(eur_lst)):
    try:
        s=str(eur_lst[i][0])
        eur_names.append(re.findall('>([^"]*)<', s))
        eur_links.append(re.findall('"([^"]*)"', s))
    except:
        pass

eur_df=pd.DataFrame(eur_names,columns=['names'])

eur_df['link']=pd.DataFrame(eur_links,columns=['link'])

eur_df.drop_duplicates(inplace=True)

eur_df.reset_index(drop=True,inplace=True)

eur_height=pd.DataFrame(index=eur_df['link'],columns=['height'])
k=0
for i in eur_df['link']:
    test=0
    while test==0:
        try:
            url = 'https://basketball.realgm.com'+i
            html = urlopen(url)
            soup = BeautifulSoup(html, features="lxml")
            test=1
        except:
            continue
        break
    s=soup.find(string=re.compile("cm\)"))
    try:
        loc=s.find('cm')
        eur_height.loc[i,'height']=s[loc-3:loc]
    except:
        eur_height.loc[i,'height']='No'
    k+=1
    print(k)

eur_df.to_csv('eur_heights.csv',index=False)
