==> 13. Colleges.ipynb <==
import numpy as np
import pandas as pd

colleges=pd.read_excel('colleges.xlsx')

college_df=pd.DataFrame()
    
k=0
j=0
for i in range(1,len(colleges)):
    college_df.loc[j,k]=colleges.loc[i,'Names']
    k+=1
    if k==5:
        k=0
        j+=1
        

college_df.rename(columns={0:'school',1:'city',2:'region',3:'conference',4:'division'},inplace=True)
